adding html css and js files for simple portfolio
This is my personal web developer portfolio built using HTML, CSS, and JavaScript. It showcases some of the projects I've created, including a Calculator App and a To-Do List, along with a summary of my skills and contact links. The portfolio is fully responsive and reflects my growing frontend development skills.